Some refactoring and fixing unclear things

+++ src/ru/fizteh/fivt/students/standy66_new/storage/strings/StringTable.java
@@ -1,6 +1,7 @@
 package ru.fizteh.fivt.students.standy66_new.storage.strings;
 
 import ru.fizteh.fivt.storage.strings.Table;
+import ru.fizteh.fivt.students.standy66_new.exceptions.CheckedExceptionCaughtException;
 import ru.fizteh.fivt.students.standy66_new.storage.FileMap;
 
 import java.io.File;
@@ -25,7 +26,7 @@ public class StringTable implements Table {
             throw new IllegalArgumentException("table directory should not point to a regular file");
         }
         this.tableDirectory = tableDirectory;
-        //noinspection CollectionWithoutInitialCapacity
+
         openedFiles = new HashMap<>();
 
         for (int i = 0; i < MAX_DIR_CHUNK; i++) {
@@ -33,10 +34,10 @@ public class StringTable implements Table {
             for (int j = 0; j < MAX_FILE_CHUNK; j++) {
                 File file = new File(dir, j + ".dat");
                 try {
-                    //noinspection resource,ObjectAllocationInLoop
+
                     openedFiles.put(file, new FileMap(file));
                 } catch (IOException e) {
-                    throw new RuntimeException("IOException occurred", e);
+                    throw new CheckedExceptionCaughtException("IOException occurred", e);
                 }
             }
         }
@@ -56,14 +57,9 @@ public class StringTable implements Table {
         if (key == null) {
             throw new IllegalArgumentException("key should not be null");
         }
-        //noinspection resource
+
         FileMap fm = getFileMapByKey(key);
-        if (fm == null) {
-            //noinspection ReturnOfNull
-            return null;
-        } else {
-            return fm.get(key);
-        }
+        return (fm == null) ? null : fm.get(key);
     }
 
     @Override
@@ -71,14 +67,9 @@ public class StringTable implements Table {
         if (key == null) {
             throw new IllegalArgumentException("key should not be null");
         }
-        //noinspection resource
+
         FileMap fm = getFileMapByKey(key);
-        if (fm == null) {
-            //noinspection ReturnOfNull
-            return null;
-        } else {
-            return fm.put(key, value);
-        }
+        return (fm == null) ? null : fm.put(key, value);
     }
 
     @Override
@@ -86,14 +77,9 @@ public class StringTable implements Table {
         if (key == null) {
             throw new IllegalArgumentException("key is null");
         }
-        //noinspection resource
+
         FileMap fm = getFileMapByKey(key);
-        if (fm == null) {
-            //noinspection ReturnOfNull
-            return null;
-        } else {
-            return fm.remove(key);
-        }
+        return (fm == null) ? null : fm.remove(key);
     }
 
     public int unsavedChangesCount() {
@@ -112,14 +98,14 @@ public class StringTable implements Table {
             try {
                 keyChangedCount += fm.commit();
             } catch (IOException e) {
-                throw new RuntimeException("IOException occurred", e);
+                throw new CheckedExceptionCaughtException("IOException occurred", e);
             }
         }
-        for (int i = 0; i < 16; i++) {
+        for (int i = 0; i < MAX_DIR_CHUNK; i++) {
             File dir = new File(tableDirectory, String.format("%d.dir", i));
-            //noinspection ConstantConditions
-            if (dir.exists() && dir.listFiles().length == 0) {
-                //noinspection ResultOfMethodCallIgnored
+
+            if (dir.exists() && (dir.listFiles() != null) && (dir.listFiles().length == 0)) {
+
                 dir.delete();
             }
         }
@@ -133,7 +119,7 @@ public class StringTable implements Table {
             try {
                 keyChangedCount += fm.rollback();
             } catch (IOException e) {
-                throw new RuntimeException("IOException occurred", e);
+                throw new CheckedExceptionCaughtException("IOException occurred", e);
             }
         }
         return keyChangedCount;
@@ -141,7 +127,7 @@ public class StringTable implements Table {
 
     @Override
     public List<String> list() {
-        //noinspection CollectionWithoutInitialCapacity
+
         return new ArrayList<>(openedFiles.values().stream().map(FileMap::keySet)
                 .reduce(new HashSet<>(), (accumulator, set) -> {
                     accumulator.addAll(set);
